feat(scraper): add support for headless mode
This commit adds support for running the scraper in headless mode, which means that the browser window will not be visible during scraping. This can be useful for scenarios where the user does not want to interact with the browser manually. The new feature has been implemented using the Playwright library and is available in all Scraper instances.